decomposes scripts into separate files, fixes the template home-card-movie.hbs
