+++ src/index.js
@@ -5,8 +5,6 @@ import popularFilmsTpl from './templates/popular-films.hbs';
 import searchFilmsTpl from './templates/home-card-movie';
 import { refs } from './js/refs';
 
-document.addEventListener('DOMContentLoaded', onHomePageLoad);
-refs.weekBtn.addEventListener('click', onWeekBtnClick);
 refs.searchMovieForm.addEventListener('submit', onSearch);
 
 const moviesApiService = new MoviesApiService();
@@ -23,21 +21,3 @@ function renderResaultsMarkup(results) {
   refs.sectionContainer.insertAdjacentHTML('beforeend', searchFilmsTpl(results));
 }
 
-function onHomePageLoad() {
-  refs.sectionContainer.innerHTML = '';
-  fetchPopularDayMovies()
-    .then(movie => renderPopularMoviesCards(movie))
-    .catch(console.log);
-}
-
-function onWeekBtnClick() {
-  refs.sectionContainer.innerHTML = '';
-  fetchPopularWeekMovies()
-    .then(movie => renderPopularMoviesCards(movie))
-    .catch(console.log);
-}
-
-function renderPopularMoviesCards(movies) {
-  const movieList = popularFilmsTpl(movies);
-  refs.sectionContainer.insertAdjacentHTML('afterbegin', movieList);
-}
