+++ src/js/fetchConst.js
@@ -1,0 +1,7 @@
+const BASE_URL = 'https://api.themoviedb.org/3';
+const API_KEY = '0425708b766634a264f7b84cc81ebcd7';
+const SEARCH_MOVIE = 'search/movie'; //получение фильма по названию
+const TRANDING_DAY = 'trending/movie/day'; //получение самых популярных фильмов за день
+const TRANDING_WEEK = 'trending/movie/week'; //получение самых популярных фильмов за неделю
+
+export { BASE_URL, API_KEY, SEARCH_MOVIE, TRANDING_DAY, TRANDING_WEEK };
