+++ src/js/fetchFunctions.js
@@ -1,0 +1,13 @@
+export function fetchPopularDayMovies() {
+    return fetch(`${BASE_URL}/${TRANDING_DAY}?api_key=${API_KEY}`)
+    .then(response => {
+      return response.json();
+    });
+  }
+
+  export function fetchPopularWeekMovies() {
+    return fetch(`${BASE_URL}/${TRANDING_WEEK}?api_key=${API_KEY}`)
+    .then(response => {
+      return response.json();
+    });
+  }
