+++ src/templates/home-card-movie.hbs
@@ -1,3 +1,4 @@
+<ul class="list movies-list">
 {{#each this}}
 <li class="card-library">
     <img class="img-card-library" src="https://image.tmdb.org/t/p/original{{poster_path}}" alt="" data-src="" />
@@ -9,3 +10,4 @@
     </div>
 </li>
 {{/each}}
+</ul>
